配置文件修改

+++ chapter05/print_triangulation.py
@@ -6,7 +6,7 @@
 @email: huijz8117@gmail.com
 @version: 0.1
 """
-rows = int(raw_input('输入列数： '))
+rows = float(raw_input('输入列数： '))
 i = j = k = 1  # 声明变量，i用于控制外层循环（图形行数），j用于控制空格的个数，k用于控制*的个数
 # 等腰直角三角形1
 print "等腰直角三角形1"
